+++ NEWS.md
@@ -1,8 +1,10 @@
 # migbirdHIP (dev version)
 
-## Major changes & new features
-
 ## Minor changes / bug fixes
+
+-   `read_hip()`: add message that reports how long it took to read in `N` number of files
+-   `writeReport()`: add more dependency checks
+-   `badRegYearMessage()`: summarize by source file because it is more informative
 
 # migbirdHIP 1.4.0